# OpenApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Guids.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/guid": {
      "post": {
        "tags": [
          "Guid"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGuidRq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGuidRq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGuidRq"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GuidInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/guid/{guid}": {
      "get": {
        "tags": [
          "Guid"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GuidInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidInfo"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Guid"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGuidMetadataRq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGuidMetadataRq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGuidMetadataRq"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GuidInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidInfo"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Guid"
        ],
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateGuidRq": {
        "required": [
          "user"
        ],
        "type": "object",
        "properties": {
          "user": {
            "minLength": 1,
            "type": "string"
          },
          "expires": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GuidInfo": {
        "type": "object",
        "properties": {
          "guid": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "type": "string",
            "nullable": true
          },
          "expires": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateGuidMetadataRq": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "nullable": true
          },
          "expires": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}